+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Root module (`mod.rs`, `lib.rs` or `main.rs`) generation for codegen (previously we were only generating children modules).
 - Cargo manifest generation (gated by `"cli"` feature).
 - `[bin]` target (CLI) for generating crates.
-- CLI generation (fancy curl for your APIs).
+- Support generating CLI (fancy curl for your APIs) which uses async/await and `runtime_tokio`.
 - `ApiClient::make_request` for sending a request and fetching a response future.
 
 ### Changed